refactor(tests): move helpers to shared location and wire PRIOR_RESOLUTIONS to Synthesizer
Move loadFile/extractSection from tests/decisions/helpers.ts to tests/helpers.ts
so the review test suite can import them without a cross-domain dependency.
Keep tests/decisions/helpers.ts as a re-export shim for backward compatibility.

Pass PRIOR_RESOLUTIONS to the Synthesizer in all three review orchestration
surfaces (code-review.md, code-review-teams.md, review:orch SKILL.md) so its
declared cross-referencing step 5 is no longer dead code.
