+++ tests/decisions/helpers.ts
@@ -1,22 +1,3 @@
-import { readFileSync } from 'fs'
-import * as path from 'path'
-
-export const ROOT = path.resolve(import.meta.dirname, '../..')
-
-export function loadFile(relPath: string): string {
-  return readFileSync(path.join(ROOT, relPath), 'utf8')
-}
-
-/**
- * Extract a named section from markdown content.
- * Returns the content from startAnchor to endAnchor (or end of string).
- * Throws loudly if either anchor is absent.
- */
-export function extractSection(content: string, startAnchor: string, endAnchor: string | null): string {
-  const start = content.indexOf(startAnchor)
-  if (start === -1) throw new Error(`Anchor not found: "${startAnchor}"`)
-  if (endAnchor === null) return content.slice(start)
-  const end = content.indexOf(endAnchor, start + startAnchor.length)
-  if (end === -1) throw new Error(`End anchor not found after "${startAnchor}": "${endAnchor}"`)
-  return content.slice(start, end)
-}
+// Re-export from shared test utilities. Kept for backward compatibility
+// with tests/decisions/helpers.test.ts which imports from './helpers'.
+export { ROOT, loadFile, extractSection } from '../helpers'
